(maint) Move PuppetDB helpers to spec_helper
 - This code is identical to what is in the PDB specs, so just copy
   it in to what will become the shared spec gem
