+++ spec/examples/running_cluster.rb
@@ -5,30 +5,6 @@ shared_examples 'a running pupperware cluster' do
 
   include Pupperware::SpecHelpers
 
-  def get_puppetdb_state
-    # make sure PDB container hasn't stopped
-    get_service_container('puppetdb', 5)
-    # now query its status endpoint
-    pdb_uri = URI::join(get_service_base_uri('puppetdb', 8080), '/status/v1/services/puppetdb-status')
-    response = Net::HTTP.get_response(pdb_uri)
-    STDOUT.puts "retrieved raw puppetdb status: #{response.body}"
-    case response
-      when Net::HTTPSuccess then
-        return JSON.parse(response.body)['state']
-      else
-        return ''
-    end
-  rescue Errno::ECONNREFUSED, Errno::ECONNRESET, EOFError => e
-    STDOUT.puts "PDB not accepting connections yet #{pdb_uri}: #{e}"
-    return ''
-  rescue JSON::ParserError
-    STDOUT.puts "Invalid JSON response: #{e}"
-    return ''
-  rescue
-    STDOUT.puts "Failure querying #{pdb_uri}: #{$!}"
-    raise
-  end
-
   def start_puppetserver
     container = get_service_container('puppet')
     status = get_container_status(container)
@@ -74,15 +50,6 @@ shared_examples 'a running pupperware cluster' do
     STDOUT.puts "cleaning cert for #{agent_name}.#{domain}"
     result = run_command("docker-compose --no-ansi exec -T puppet puppetserver ca clean --certname #{agent_name}.#{domain}")
     return result[:status].exitstatus
-  end
-
-  def wait_on_puppetdb_status(seconds = 240)
-    # since pdb doesn't have a proper healthcheck yet, this could spin forever
-    # add a timeout so it eventually returns.
-    return retry_block_up_to_timeout(seconds) do
-      get_puppetdb_state() == 'running' ? 'running' :
-        raise('puppetdb never entered running state')
-    end
   end
 
   it 'should start all of the cluster services' do
